requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
opencv-python>=4.12.0
opencv-contrib-python>=4.12.0
Pillow>=10.0.0
numpy>=2.0.0
requests>=2.28.0 